requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
pydantic-settings==2.1.0
pyjwt==2.8.0
python-multipart==0.0.6
structlog==23.2.0
slowapi==0.1.9
redis==5.0.1
prometheus-client==0.19.0
cryptography==41.0.7
